requirements.txt
# Production dependencies
fastmcp
requests
python-dotenv
fastapi
uvicorn
pyyaml